Do NOT delete video of test, even if test is successful IF environment variable - KEEP_ALL_VIDEOS is set to true

+++ src/main/java/com/appium/manager/TestLogger.java
@@ -136,13 +136,15 @@ class TestLogger extends Helpers {
     }
 
     private void deleteSuccessVideos(ITestResult result, String className) {
-        if (result.isSuccess()) {
+        if (result.isSuccess()
+                && (null != System.getenv("KEEP_ALL_VIDEOS"))
+                && !(System.getenv("KEEP_ALL_VIDEOS").equalsIgnoreCase("true"))) {
             File videoFile = new File(System.getProperty("user.dir")
                     + FileLocations.ANDROID_SCREENSHOTS_DIRECTORY
                     + AppiumDeviceManager.getAppiumDevice().getDevice().getUdid() + "/"
                     + className + "/" + result.getMethod().getMethodName()
                     + "/" + result.getMethod().getMethodName() + ".mp4");
-            System.out.println(videoFile);
+            System.out.println(String.format("Deleting video file - %s, if it exists", videoFile));
             if (videoFile.exists()) {
                 videoFile.delete();
             }
